+++ game/src/game/dao/IOptionDao.java
@@ -3,17 +3,15 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package game.Models;
+package game.dao;
 
-import game.entity.QuizDto;
+import game.entity.OptionDto;
 import java.util.List;
 
 /**
- * The Object for writing to and reading  from a quiz file.
+ * The Object for writing to and reading from a option file.
  * @author Masaomi
  */
-public interface IQuizDao {
-    
-    List<QuizDto> getByLevel(int level);
-    
+public interface IOptionDao {
+    List<OptionDto> getById(int id);
 }
