+++ game/src/game/dao/UserDao.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package game.Models;
+package game.dao;
 
 import game.entity.Player;
 import game.util.Const;
